pyproject.toml
[tool.poetry]
name = "hfc-backend"
version = "0.1.0"
description = "Family website API for the Hoar family christmas drawing\u001b[D\u001b[D"
authors = ["Robert Townley <me@roberttownley.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
pylint-django = "^2.4.4"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
pylint-django = "^2.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
